pull from refactor_getbh
