+++ magpylib/_src/defaults/defaults_utility.py
@@ -5,19 +5,8 @@ from copy import deepcopy
 from magpylib._src.defaults.defaults_values import DEFAULTS
 
 
-SUPPORTED_PLOTTING_BACKENDS = ("matplotlib", "plotly", "matplotlib_old")
+SUPPORTED_PLOTTING_BACKENDS = ("matplotlib", "plotly")
 
-MAGPYLIB_FAMILIES = {
-    "Line": ("current",),
-    "Loop": ("current",),
-    "Cuboid": ("magnet",),
-    "Cylinder": ("magnet",),
-    "Sphere": ("magnet",),
-    "CylinderSegment": ("magnet",),
-    "Sensor": ("sensor",),
-    "Dipole": ("dipole",),
-    "Marker": ("markers",),
-}
 
 SYMBOLS_MATPLOTLIB_TO_PLOTLY = {
     ".": "circle",
