+++ magpylib/_src/fields/__init__.py
@@ -3,4 +3,4 @@
 __all__ = ["getB", "getH"]
 
 # create interface to outside of package
-from magpylib._src.fields.field_wrap_BH_level3 import getB, getH
+from magpylib._src.fields.field_wrap_BH import getB, getH
