+++ magpylib/_src/obj_classes/class_BaseGetBH.py
@@ -1,7 +1,7 @@
 """BaseGetBHsimple class code
 DOCSTRINGS V4 READY
 """
-from magpylib._src.fields.field_wrap_BH_level2 import getBH_level2
+from magpylib._src.fields.field_wrap_BH import getBH_level2
 from magpylib._src.utility import format_star_input
