+++ magpylib/_src/obj_classes/class_Sensor.py
@@ -1,13 +1,15 @@
 """Sensor class code
 DOCSTRINGS V4 READY
 """
-from magpylib._src.fields.field_wrap_BH_level2 import getBH_level2
+from magpylib._src.fields.field_wrap_BH import getBH_level2
 from magpylib._src.input_checks import check_format_input_vector
 from magpylib._src.obj_classes.class_BaseDisplayRepr import BaseDisplayRepr
 from magpylib._src.obj_classes.class_BaseGeo import BaseGeo
 from magpylib._src.utility import format_star_input
+from magpylib._src.utility import Registered
 
 
+@Registered(kind="sensor", family="sensor")
 class Sensor(BaseGeo, BaseDisplayRepr):
     """Magnetic field sensor.
 
@@ -85,7 +87,6 @@ class Sensor(BaseGeo, BaseDisplayRepr):
 
         # instance attributes
         self.pixel = pixel
-        self._object_type = "Sensor"
 
         # init inheritance
         BaseGeo.__init__(self, position, orientation, style=style, **kwargs)
