+++ magpylib/_src/obj_classes/class_magnet_Cuboid.py
@@ -1,20 +1,29 @@
-"""Magnet Sphere class code
+"""Magnet Cuboid class code
 DOCSTRINGS V4 READY
 """
-from magpylib._src.input_checks import check_format_input_scalar
+from magpylib._src.fields.field_BH_cuboid import magnet_cuboid_field
+from magpylib._src.input_checks import check_format_input_vector
 from magpylib._src.obj_classes.class_BaseDisplayRepr import BaseDisplayRepr
 from magpylib._src.obj_classes.class_BaseExcitations import BaseHomMag
 from magpylib._src.obj_classes.class_BaseGeo import BaseGeo
 from magpylib._src.obj_classes.class_BaseGetBH import BaseGetBH
+from magpylib._src.utility import Registered
 
 
-class Sphere(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
-    """Spherical magnet with homogeneous magnetization.
+@Registered(
+    kind="source",
+    family="magnet",
+    field_func=magnet_cuboid_field,
+    source_kwargs_ndim={"magnetization": 2, "dimension": 2},
+)
+class Cuboid(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
+    """Cuboid magnet with homogeneous magnetization.
 
     Can be used as `sources` input for magnetic field computation.
 
-    When `position=(0,0,0)` and `orientation=None` the sphere center is located
-    in the origin of the global coordinate system.
+    When `position=(0,0,0)` and `orientation=None` the Cuboid sides are parallel
+    to the global coordinate basis vectors and the geometric center of the Cuboid
+    is located in the origin.
 
     Parameters
     ----------
@@ -22,8 +31,8 @@ class Sphere(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
         Magnetization vector (mu0*M, remanence field) in units of [mT] given in
         the local object coordinates (rotates with object).
 
-    diameter: float, default=`None`
-        Diameter of the sphere in units of [mm].
+    dimension: array_like, shape (3,), default=`None`
+        Length of the cuboid sides [a,b,c] in units of [mm].
 
     position: array_like, shape (3,) or (m,3), default=`(0,0,0)`
         Object position(s) in the global coordinates in units of [mm]. For m>1, the
@@ -41,50 +50,49 @@ class Sphere(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
         Object style inputs must be in dictionary form, e.g. `{'color':'red'}` or
         using style underscore magic, e.g. `style_color='red'`.
 
-
     Returns
     -------
-    magnet source: `Sphere` object
+    magnet source: `Cuboid` object
 
     Examples
     --------
-    `Sphere` objects are magnetic field sources. In this example we compute the H-field [kA/m]
-    of a spherical magnet with magnetization (100,200,300) in units of [mT] and diameter
-    of 1 [mm] at the observer position (1,1,1) given in units of [mm]:
+    `Cuboid` magnets are magnetic field sources. Below we compute the H-field [kA/m] of a
+    cubical magnet with magnetization (100,200,300) in units of [mT] and 1 [mm] sides
+    at the observer position (1,1,1) given in units of [mm]:
 
     >>> import magpylib as magpy
-    >>> src = magpy.magnet.Sphere(magnetization=(100,200,300), diameter=1)
+    >>> src = magpy.magnet.Cuboid(magnetization=(100,200,300), dimension=(1,1,1))
     >>> H = src.getH((1,1,1))
     >>> print(H)
-    [3.19056074 2.55244859 1.91433644]
+    [6.21116976 4.9689358  3.72670185]
 
     We rotate the source object, and compute the B-field, this time at a set of observer positions:
 
     >>> src.rotate_from_angax(45, 'x')
-    Sphere(id=...)
+    Cuboid(id=...)
     >>> B = src.getB([(1,1,1), (2,2,2), (3,3,3)])
     >>> print(B)
-    [[2.26804606 3.63693295 0.23486386]
-     [0.28350576 0.45461662 0.02935798]
-     [0.08400171 0.13470122 0.00869866]]
+    [[4.30496934 6.9363475  0.50728577]
+     [0.54127889 0.86827283 0.05653357]
+     [0.1604214  0.25726266 0.01664045]]
 
     The same result is obtained when the rotated source moves along a path away from an
-    observer at position (1,1,1). This time we use a `Sensor` object as observer.
+    observer at position (1,1,1). Here we use a `Sensor` object as observer.
 
-    >>> src.move([(-1,-1,-1), (-2,-2,-2)])
-    Sphere(id=...)
     >>> sens = magpy.Sensor(position=(1,1,1))
+    >>> src.move([(-1,-1,-1), (-2,-2,-2)])
+    Cuboid(id=...)
     >>> B = src.getB(sens)
     >>> print(B)
-    [[2.26804606 3.63693295 0.23486386]
-     [0.28350576 0.45461662 0.02935798]
-     [0.08400171 0.13470122 0.00869866]]
+    [[4.30496934 6.9363475  0.50728577]
+     [0.54127889 0.86827283 0.05653357]
+     [0.1604214  0.25726266 0.01664045]]
     """
 
     def __init__(
         self,
         magnetization=None,
-        diameter=None,
+        dimension=None,
         position=(0, 0, 0),
         orientation=None,
         style=None,
@@ -92,8 +100,7 @@ class Sphere(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
     ):
 
         # instance attributes
-        self.diameter = diameter
-        self._object_type = "Sphere"
+        self.dimension = dimension
 
         # init inheritance
         BaseGeo.__init__(self, position, orientation, style=style, **kwargs)
@@ -102,17 +109,19 @@ class Sphere(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
 
     # property getters and setters
     @property
-    def diameter(self):
-        """Diameter of the sphere in units of [mm]."""
-        return self._diameter
+    def dimension(self):
+        """Length of the cuboid sides [a,b,c] in units of [mm]."""
+        return self._dimension
 
-    @diameter.setter
-    def diameter(self, dia):
-        """Set Sphere diameter, float, [mm]."""
-        self._diameter = check_format_input_scalar(
-            dia,
-            sig_name="diameter",
-            sig_type="`None` or a positive number (int, float)",
+    @dimension.setter
+    def dimension(self, dim):
+        """Set Cuboid dimension (a,b,c), shape (3,), [mm]."""
+        self._dimension = check_format_input_vector(
+            dim,
+            dims=(1,),
+            shape_m1=3,
+            sig_name="Cuboid.dimension",
+            sig_type="array_like (list, tuple, ndarray) of shape (3,) with positive values",
             allow_None=True,
-            forbid_negative=True,
+            forbid_negative0=True,
         )
