+++ magpylib/_src/obj_classes/class_magnet_Cylinder.py
@@ -1,21 +1,29 @@
-"""Magnet Cuboid class code
+"""Magnet Cylinder class code
 DOCSTRINGS V4 READY
 """
+from magpylib._src.fields.field_BH_cylinder_segment import magnet_cylinder_field
 from magpylib._src.input_checks import check_format_input_vector
 from magpylib._src.obj_classes.class_BaseDisplayRepr import BaseDisplayRepr
 from magpylib._src.obj_classes.class_BaseExcitations import BaseHomMag
 from magpylib._src.obj_classes.class_BaseGeo import BaseGeo
 from magpylib._src.obj_classes.class_BaseGetBH import BaseGetBH
+from magpylib._src.utility import Registered
 
 
-class Cuboid(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
-    """Cuboid magnet with homogeneous magnetization.
+@Registered(
+    kind="source",
+    family="magnet",
+    field_func=magnet_cylinder_field,
+    source_kwargs_ndim={"magnetization": 2, "dimension": 2},
+)
+class Cylinder(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
+    """Cylinder magnet with homogeneous magnetization.
 
     Can be used as `sources` input for magnetic field computation.
 
-    When `position=(0,0,0)` and `orientation=None` the Cuboid sides are parallel
-    to the global coordinate basis vectors and the geometric center of the Cuboid
-    is located in the origin.
+    When `position=(0,0,0)` and `orientation=None` the geometric center of the
+    cylinder lies in the origin of the global coordinate system and
+    the cylinder axis conincides with the global z-axis.
 
     Parameters
     ----------
@@ -23,8 +31,8 @@ class Cuboid(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
         Magnetization vector (mu0*M, remanence field) in units of [mT] given in
         the local object coordinates (rotates with object).
 
-    dimension: array_like, shape (3,), default=`None`
-        Length of the cuboid sides [a,b,c] in units of [mm].
+    dimension: array_like, shape (2,), default=`None`
+        Dimension (d,h) denote diameter and height of the cylinder in units of [mm].
 
     position: array_like, shape (3,) or (m,3), default=`(0,0,0)`
         Object position(s) in the global coordinates in units of [mm]. For m>1, the
@@ -44,41 +52,41 @@ class Cuboid(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
 
     Returns
     -------
-    magnet source: `Cuboid` object
+    magnet source: `Cylinder`
 
     Examples
     --------
-    `Cuboid` magnets are magnetic field sources. Below we compute the H-field [kA/m] of a
-    cubical magnet with magnetization (100,200,300) in units of [mT] and 1 [mm] sides
+    `Cylinder` magnets are magnetic field sources. Below we compute the H-field [kA/m] of a
+    cylinder magnet with magnetization (100,200,300) in units of [mT] and 1 [mm] diameter and height
     at the observer position (1,1,1) given in units of [mm]:
 
     >>> import magpylib as magpy
-    >>> src = magpy.magnet.Cuboid(magnetization=(100,200,300), dimension=(1,1,1))
+    >>> src = magpy.magnet.Cylinder(magnetization=(100,200,300), dimension=(1,1))
     >>> H = src.getH((1,1,1))
     >>> print(H)
-    [6.21116976 4.9689358  3.72670185]
+    [4.84991343 3.88317816 2.73973202]
 
     We rotate the source object, and compute the B-field, this time at a set of observer positions:
 
     >>> src.rotate_from_angax(45, 'x')
-    Cuboid(id=...)
+    Cylinder(id=...)
     >>> B = src.getB([(1,1,1), (2,2,2), (3,3,3)])
     >>> print(B)
-    [[4.30496934 6.9363475  0.50728577]
-     [0.54127889 0.86827283 0.05653357]
-     [0.1604214  0.25726266 0.01664045]]
+    [[3.31419501 5.26683023 0.37767015]
+     [0.42298405 0.67710536 0.04464932]
+     [0.12571523 0.20144503 0.01312389]]
 
     The same result is obtained when the rotated source moves along a path away from an
     observer at position (1,1,1). Here we use a `Sensor` object as observer.
 
     >>> sens = magpy.Sensor(position=(1,1,1))
     >>> src.move([(-1,-1,-1), (-2,-2,-2)])
-    Cuboid(id=...)
+    Cylinder(id=...)
     >>> B = src.getB(sens)
     >>> print(B)
-    [[4.30496934 6.9363475  0.50728577]
-     [0.54127889 0.86827283 0.05653357]
-     [0.1604214  0.25726266 0.01664045]]
+    [[3.31419501 5.26683023 0.37767015]
+     [0.42298405 0.67710536 0.04464932]
+     [0.12571523 0.20144503 0.01312389]]
     """
 
     def __init__(
@@ -93,7 +101,6 @@ class Cuboid(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
 
         # instance attributes
         self.dimension = dimension
-        self._object_type = "Cuboid"
 
         # init inheritance
         BaseGeo.__init__(self, position, orientation, style=style, **kwargs)
@@ -103,18 +110,18 @@ class Cuboid(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
     # property getters and setters
     @property
     def dimension(self):
-        """Length of the cuboid sides [a,b,c] in units of [mm]."""
+        """Dimension (d,h) denote diameter and height of the cylinder in units of [mm]."""
         return self._dimension
 
     @dimension.setter
     def dimension(self, dim):
-        """Set Cuboid dimension (a,b,c), shape (3,), [mm]."""
+        """Set Cylinder dimension (d,h) in units of [mm]."""
         self._dimension = check_format_input_vector(
             dim,
             dims=(1,),
-            shape_m1=3,
-            sig_name="Cuboid.dimension",
-            sig_type="array_like (list, tuple, ndarray) of shape (3,) with positive values",
+            shape_m1=2,
+            sig_name="Cylinder.dimension",
+            sig_type="array_like (list, tuple, ndarray) with shape (2,) with positive values",
             allow_None=True,
             forbid_negative0=True,
         )
