+++ magpylib/_src/obj_classes/class_magnet_Sphere.py
@@ -1,21 +1,28 @@
-"""Magnet Cylinder class code
+"""Magnet Sphere class code
 DOCSTRINGS V4 READY
 """
-from magpylib._src.input_checks import check_format_input_vector
+from magpylib._src.fields.field_BH_sphere import magnet_sphere_field
+from magpylib._src.input_checks import check_format_input_scalar
 from magpylib._src.obj_classes.class_BaseDisplayRepr import BaseDisplayRepr
 from magpylib._src.obj_classes.class_BaseExcitations import BaseHomMag
 from magpylib._src.obj_classes.class_BaseGeo import BaseGeo
 from magpylib._src.obj_classes.class_BaseGetBH import BaseGetBH
+from magpylib._src.utility import Registered
 
 
-class Cylinder(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
-    """Cylinder magnet with homogeneous magnetization.
+@Registered(
+    kind="source",
+    family="magnet",
+    field_func=magnet_sphere_field,
+    source_kwargs_ndim={"magnetization": 2, "diameter": 1},
+)
+class Sphere(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
+    """Spherical magnet with homogeneous magnetization.
 
     Can be used as `sources` input for magnetic field computation.
 
-    When `position=(0,0,0)` and `orientation=None` the geometric center of the
-    cylinder lies in the origin of the global coordinate system and
-    the cylinder axis conincides with the global z-axis.
+    When `position=(0,0,0)` and `orientation=None` the sphere center is located
+    in the origin of the global coordinate system.
 
     Parameters
     ----------
@@ -23,8 +30,8 @@ class Cylinder(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
         Magnetization vector (mu0*M, remanence field) in units of [mT] given in
         the local object coordinates (rotates with object).
 
-    dimension: array_like, shape (2,), default=`None`
-        Dimension (d,h) denote diameter and height of the cylinder in units of [mm].
+    diameter: float, default=`None`
+        Diameter of the sphere in units of [mm].
 
     position: array_like, shape (3,) or (m,3), default=`(0,0,0)`
         Object position(s) in the global coordinates in units of [mm]. For m>1, the
@@ -42,49 +49,50 @@ class Cylinder(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
         Object style inputs must be in dictionary form, e.g. `{'color':'red'}` or
         using style underscore magic, e.g. `style_color='red'`.
 
+
     Returns
     -------
-    magnet source: `Cylinder`
+    magnet source: `Sphere` object
 
     Examples
     --------
-    `Cylinder` magnets are magnetic field sources. Below we compute the H-field [kA/m] of a
-    cylinder magnet with magnetization (100,200,300) in units of [mT] and 1 [mm] diameter and height
-    at the observer position (1,1,1) given in units of [mm]:
+    `Sphere` objects are magnetic field sources. In this example we compute the H-field [kA/m]
+    of a spherical magnet with magnetization (100,200,300) in units of [mT] and diameter
+    of 1 [mm] at the observer position (1,1,1) given in units of [mm]:
 
     >>> import magpylib as magpy
-    >>> src = magpy.magnet.Cylinder(magnetization=(100,200,300), dimension=(1,1))
+    >>> src = magpy.magnet.Sphere(magnetization=(100,200,300), diameter=1)
     >>> H = src.getH((1,1,1))
     >>> print(H)
-    [4.84991343 3.88317816 2.73973202]
+    [3.19056074 2.55244859 1.91433644]
 
     We rotate the source object, and compute the B-field, this time at a set of observer positions:
 
     >>> src.rotate_from_angax(45, 'x')
-    Cylinder(id=...)
+    Sphere(id=...)
     >>> B = src.getB([(1,1,1), (2,2,2), (3,3,3)])
     >>> print(B)
-    [[3.31419501 5.26683023 0.37767015]
-     [0.42298405 0.67710536 0.04464932]
-     [0.12571523 0.20144503 0.01312389]]
+    [[2.26804606 3.63693295 0.23486386]
+     [0.28350576 0.45461662 0.02935798]
+     [0.08400171 0.13470122 0.00869866]]
 
     The same result is obtained when the rotated source moves along a path away from an
-    observer at position (1,1,1). Here we use a `Sensor` object as observer.
+    observer at position (1,1,1). This time we use a `Sensor` object as observer.
 
-    >>> sens = magpy.Sensor(position=(1,1,1))
     >>> src.move([(-1,-1,-1), (-2,-2,-2)])
-    Cylinder(id=...)
+    Sphere(id=...)
+    >>> sens = magpy.Sensor(position=(1,1,1))
     >>> B = src.getB(sens)
     >>> print(B)
-    [[3.31419501 5.26683023 0.37767015]
-     [0.42298405 0.67710536 0.04464932]
-     [0.12571523 0.20144503 0.01312389]]
+    [[2.26804606 3.63693295 0.23486386]
+     [0.28350576 0.45461662 0.02935798]
+     [0.08400171 0.13470122 0.00869866]]
     """
 
     def __init__(
         self,
         magnetization=None,
-        dimension=None,
+        diameter=None,
         position=(0, 0, 0),
         orientation=None,
         style=None,
@@ -92,8 +100,7 @@ class Cylinder(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
     ):
 
         # instance attributes
-        self.dimension = dimension
-        self._object_type = "Cylinder"
+        self.diameter = diameter
 
         # init inheritance
         BaseGeo.__init__(self, position, orientation, style=style, **kwargs)
@@ -102,19 +109,17 @@ class Cylinder(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseHomMag):
 
     # property getters and setters
     @property
-    def dimension(self):
-        """Dimension (d,h) denote diameter and height of the cylinder in units of [mm]."""
-        return self._dimension
+    def diameter(self):
+        """Diameter of the sphere in units of [mm]."""
+        return self._diameter
 
-    @dimension.setter
-    def dimension(self, dim):
-        """Set Cylinder dimension (d,h) in units of [mm]."""
-        self._dimension = check_format_input_vector(
-            dim,
-            dims=(1,),
-            shape_m1=2,
-            sig_name="Cylinder.dimension",
-            sig_type="array_like (list, tuple, ndarray) with shape (2,) with positive values",
+    @diameter.setter
+    def diameter(self, dia):
+        """Set Sphere diameter, float, [mm]."""
+        self._diameter = check_format_input_scalar(
+            dia,
+            sig_name="diameter",
+            sig_type="`None` or a positive number (int, float)",
             allow_None=True,
-            forbid_negative0=True,
+            forbid_negative=True,
         )
