+++ magpylib/_src/obj_classes/class_misc_CustomSource.py
@@ -3,8 +3,10 @@ from magpylib._src.input_checks import validate_field_func
 from magpylib._src.obj_classes.class_BaseDisplayRepr import BaseDisplayRepr
 from magpylib._src.obj_classes.class_BaseGeo import BaseGeo
 from magpylib._src.obj_classes.class_BaseGetBH import BaseGetBH
+from magpylib._src.utility import Registered
 
 
+@Registered(kind="source", family="misc", field_func=None)
 class CustomSource(BaseGeo, BaseDisplayRepr, BaseGetBH):
     """User-defined custom source.
 
@@ -91,7 +93,6 @@ class CustomSource(BaseGeo, BaseDisplayRepr, BaseGetBH):
     ):
         # instance attributes
         self.field_func = field_func
-        self._object_type = "CustomSource"
 
         # init inheritance
         BaseGeo.__init__(self, position, orientation, style=style, **kwargs)
