+++ magpylib/magnet/__init__.py
@@ -5,7 +5,7 @@ The `magpylib.magnet` subpackage contains all magnet classes.
 __all__ = ["Cuboid", "Cylinder", "Sphere", "CylinderSegment"]
 
 
-from magpylib._src.obj_classes.class_mag_Cuboid import Cuboid
-from magpylib._src.obj_classes.class_mag_Cylinder import Cylinder
-from magpylib._src.obj_classes.class_mag_Sphere import Sphere
-from magpylib._src.obj_classes.class_mag_CylinderSegment import CylinderSegment
+from magpylib._src.obj_classes.class_magnet_Cuboid import Cuboid
+from magpylib._src.obj_classes.class_magnet_Cylinder import Cylinder
+from magpylib._src.obj_classes.class_magnet_Sphere import Sphere
+from magpylib._src.obj_classes.class_magnet_CylinderSegment import CylinderSegment
