+++ magpylib/misc/__init__.py
@@ -5,4 +5,4 @@ The `magpylib.misc` sub-package contains miscellaneous source objects.
 __all__ = ["Dipole", "CustomSource"]
 
 from magpylib._src.obj_classes.class_misc_Dipole import Dipole
-from magpylib._src.obj_classes.class_misc_Custom import CustomSource
+from magpylib._src.obj_classes.class_misc_CustomSource import CustomSource
