+++ tests/test_exceptions.py
@@ -1,17 +1,17 @@
 import unittest
 
 import numpy as np
-from scipy.spatial.transform import Rotation as R
 
 import magpylib as magpy
 from magpylib._src.exceptions import MagpylibBadUserInput
 from magpylib._src.exceptions import MagpylibInternalError
-from magpylib._src.fields.field_wrap_BH_level1 import getBH_level1
-from magpylib._src.fields.field_wrap_BH_level2 import getBH_level2
+from magpylib._src.fields.field_wrap_BH import getBH_level2
 from magpylib._src.input_checks import check_format_input_observers
 from magpylib._src.utility import format_obj_input
 from magpylib._src.utility import format_src_inputs
 from magpylib._src.utility import test_path_format as tpf
+
+GETBH_KWARGS = {"sumup": False, "squeeze": True, "pixel_agg": None, "output": "ndarray"}
 
 
 def getBHv_unknown_source_type():
@@ -22,26 +22,8 @@ def getBHv_unknown_source_type():
         magnetization=(1, 0, 0),
         dimension=(0, 2, 1, 0, 360),
         position=(0, 0, -0.5),
-        sumup=False,
-        squeeze=True,
-        pixel_agg=None,
-        output="ndarray",
         field="B",
-    )
-
-
-def getBH_level1_internal_error():
-    """bad source_type input should not happen"""
-    x = np.array([(1, 2, 3)])
-    rot = R.from_quat((0, 0, 0, 1))
-    getBH_level1(
-        field="B",
-        source_type="woot",
-        magnetization=x,
-        dimension=x,
-        observers=x,
-        position=x,
-        orientation=rot,
+        **GETBH_KWARGS
     )
 
 
@@ -75,7 +57,7 @@ def getBH_level2_internal_error1():
     # pylint: disable=protected-access
     sens = magpy.Sensor()
     x = np.zeros((10, 3))
-    magpy._src.fields.field_wrap_BH_level2.get_src_dict([sens], 10, 10, x)
+    magpy._src.fields.field_wrap_BH.get_src_dict([sens], 10, 10, x)
 
 
 # getBHv missing inputs ------------------------------------------------------
@@ -83,74 +65,35 @@ def getBHv_missing_input1():
     """missing field"""
     x = np.array([(1, 2, 3)])
     getBH_level2(
-        sources="Cuboid",
-        observers=x,
-        magnetization=x,
-        dimension=x,
-        sumup=False,
-        squeeze=True,
-        pixel_agg=None,
-        output="ndarray",
+        sources="Cuboid", observers=x, magnetization=x, dimension=x, **GETBH_KWARGS
     )
 
 
 def getBHv_missing_input2():
     """missing source_type"""
     x = np.array([(1, 2, 3)])
-    getBH_level2(
-        observers=x,
-        field="B",
-        magnetization=x,
-        dimension=x,
-        sumup=False,
-        squeeze=True,
-        pixel_agg=None,
-        output="ndarray",
-    )
+    getBH_level2(observers=x, field="B", magnetization=x, dimension=x, **GETBH_KWARGS)
 
 
 def getBHv_missing_input3():
     """missing observer"""
     x = np.array([(1, 2, 3)])
     getBH_level2(
-        sources="Cuboid",
-        field="B",
-        magnetization=x,
-        dimension=x,
-        sumup=False,
-        squeeze=True,
-        pixel_agg=None,
-        output="ndarray",
+        sources="Cuboid", field="B", magnetization=x, dimension=x, **GETBH_KWARGS
     )
 
 
 def getBHv_missing_input4_cuboid():
     """missing Cuboid mag"""
     x = np.array([(1, 2, 3)])
-    getBH_level2(
-        sources="Cuboid",
-        observers=x,
-        field="B",
-        dimension=x,
-        sumup=False,
-        squeeze=True,
-        pixel_agg=None,
-        output="ndarray",
-    )
+    getBH_level2(sources="Cuboid", observers=x, field="B", dimension=x, **GETBH_KWARGS)
 
 
 def getBHv_missing_input5_cuboid():
     """missing Cuboid dim"""
     x = np.array([(1, 2, 3)])
     getBH_level2(
-        sources="Cuboid",
-        observers=x,
-        field="B",
-        magnetization=x,
-        sumup=False,
-        squeeze=True,
-        pixel_agg=None,
-        output="ndarray",
+        sources="Cuboid", observers=x, field="B", magnetization=x, **GETBH_KWARGS
     )
 
 
@@ -159,14 +102,7 @@ def getBHv_missing_input4_cyl():
     x = np.array([(1, 2, 3)])
     y = np.array([(1, 2)])
     getBH_level2(
-        sources="Cylinder",
-        observers=x,
-        field="B",
-        dimension=y,
-        sumup=False,
-        squeeze=True,
-        pixel_agg=None,
-        output="ndarray",
+        sources="Cylinder", observers=x, field="B", dimension=y, **GETBH_KWARGS
     )
 
 
@@ -174,44 +110,21 @@ def getBHv_missing_input5_cyl():
     """missing Cylinder dim"""
     x = np.array([(1, 2, 3)])
     getBH_level2(
-        sources="Cylinder",
-        observers=x,
-        field="B",
-        magnetization=x,
-        sumup=False,
-        squeeze=True,
-        pixel_agg=None,
-        output="ndarray",
+        sources="Cylinder", observers=x, field="B", magnetization=x, **GETBH_KWARGS
     )
 
 
 def getBHv_missing_input4_sphere():
     """missing Sphere mag"""
     x = np.array([(1, 2, 3)])
-    getBH_level2(
-        sources="Sphere",
-        observers=x,
-        field="B",
-        dimension=1,
-        sumup=False,
-        squeeze=True,
-        pixel_agg=None,
-        output="ndarray",
-    )
+    getBH_level2(sources="Sphere", observers=x, field="B", dimension=1, **GETBH_KWARGS)
 
 
 def getBHv_missing_input5_sphere():
     """missing Sphere dim"""
     x = np.array([(1, 2, 3)])
     getBH_level2(
-        sources="Sphere",
-        observers=x,
-        field="B",
-        magnetization=x,
-        sumup=False,
-        squeeze=True,
-        pixel_agg=None,
-        output="ndarray",
+        sources="Sphere", observers=x, field="B", magnetization=x, **GETBH_KWARGS
     )
 
 
@@ -226,10 +139,7 @@ def getBHv_bad_input1():
         field="B",
         magnetization=x2,
         dimension=x,
-        sumup=False,
-        squeeze=True,
-        pixel_agg=None,
-        output="ndarray",
+        **GETBH_KWARGS
     )
 
 
@@ -242,10 +152,7 @@ def getBHv_bad_input2():
         field="B",
         magnetization=x,
         dimension=x,
-        sumup=False,
-        squeeze=True,
-        pixel_agg=None,
-        output="ndarray",
+        **GETBH_KWARGS
     )
 
 
@@ -259,10 +166,7 @@ def getBHv_bad_input3():
         field="B",
         magnetization=x,
         dimension=x,
-        sumup=False,
-        squeeze=True,
-        pixel_agg=None,
-        output="ndarray",
+        **GETBH_KWARGS
     )
 
 
@@ -378,10 +282,6 @@ class TestExceptions(unittest.TestCase):
         self.assertRaises(MagpylibBadUserInput, getBHv_bad_input2)
         self.assertRaises(MagpylibBadUserInput, getBHv_bad_input3)
         self.assertRaises(MagpylibBadUserInput, getBHv_unknown_source_type)
-
-    def test_except_getBH_lev1(self):
-        """getBH_level1 exception testing"""
-        self.assertRaises(MagpylibInternalError, getBH_level1_internal_error)
 
     def test_except_getBH_lev2(self):
         """getBH_level2 exception testing"""
