+++ tests/test_field_functions.py
@@ -298,26 +298,29 @@ def test_field_line():
 
 def test_field_line_from_vert():
     """test the Line field from vertex input"""
-    p = np.array([(1, 2, 2), (1, 2, 3), (-1, 0, -3)])
-    curr = np.array([1, 5, -3])
+    observers = np.array([(1, 2, 2), (1, 2, 3), (-1, 0, -3)])
+    current = np.array([1, 5, -3])
 
-    vert1 = np.array(
-        [(0, 0, 0), (1, 1, 1), (2, 2, 2), (3, 3, 3), (1, 2, 3), (-3, 4, -5)]
+    vertices = np.array(
+        [
+            np.array(
+                [(0, 0, 0), (1, 1, 1), (2, 2, 2), (3, 3, 3), (1, 2, 3), (-3, 4, -5)]
+            ),
+            np.array([(0, 0, 0), (3, 3, 3), (-3, 4, -5)]),
+            np.array([(1, 2, 3), (-2, -3, 3), (3, 2, 1), (3, 3, 3)]),
+        ],
+        dtype="object",
     )
-    vert2 = np.array([(0, 0, 0), (3, 3, 3), (-3, 4, -5)])
-    vert3 = np.array([(1, 2, 3), (-2, -3, 3), (3, 2, 1), (3, 3, 3)])
 
-    pos_tiled = np.tile(p, (3, 1))
-    B_vert = current_vertices_field("B", pos_tiled, curr, [vert1, vert2, vert3])
+    B_vert = current_vertices_field("B", observers, current, vertices)
 
     B = []
-    for i, vert in enumerate([vert1, vert2, vert3]):
-        for pos in p:
-            p1 = vert[:-1]
-            p2 = vert[1:]
-            po = np.array([pos] * (len(vert) - 1))
-            cu = np.array([curr[i]] * (len(vert) - 1))
-            B += [np.sum(current_line_field("B", po, cu, p1, p2), axis=0)]
+    for obs, vert, curr in zip(observers, vertices, current):
+        p1 = vert[:-1]
+        p2 = vert[1:]
+        po = np.array([obs] * (len(vert) - 1))
+        cu = np.array([curr] * (len(vert) - 1))
+        B += [np.sum(current_line_field("B", po, cu, p1, p2), axis=0)]
     B = np.array(B)
 
     assert_allclose(B_vert, B)
