+++ tests/test_obj_BaseGeo.py
@@ -417,8 +417,8 @@ def test_describe():
 
     test = (
         "<pre>Cuboid(id=REGEX, label='x1')<br>  • parent: None <br>  • "
-        + "position: [0. 0. 0.] mm<br>  • orientation: [0. 0. 0.] degrees<br>  • "
-        + "dimension: None mm<br>  • magnetization: None mT</pre>"
+        "position: [0. 0. 0.] mm<br>  • orientation: [0. 0. 0.] degrees<br>  • "
+        "dimension: None mm<br>  • magnetization: None mT<br>  • family: magnet </pre>"
     )
     rep = x1._repr_html_()
     rep = re.sub("id=[0-9]*[0-9]", "id=REGEX", rep)
@@ -432,6 +432,7 @@ def test_describe():
         "  • orientation: [0. 0. 0.] degrees",
         "  • dimension: None mm",
         "  • magnetization: None mT",
+        "  • family: magnet ",  # INVISIBLE SPACE
     ]
     desc = x1.describe(return_string=True)
     desc = re.sub("id=*[0-9]*[0-9]", "id=REGEX", desc)
@@ -444,6 +445,7 @@ def test_describe():
         "  • orientation: [0. 0. 0.] degrees",
         "  • dimension: [1. 3.] mm",
         "  • magnetization: [2. 3. 4.] mT",
+        "  • family: magnet ",  # INVISIBLE SPACE
     ]
     desc = x2.describe(return_string=True)
     desc = re.sub("id=*[0-9]*[0-9]", "id=REGEX", desc)
@@ -455,6 +457,7 @@ def test_describe():
         "  • path length: 3",
         "  • position (last): [1. 2. 3.] mm",
         "  • orientation (last): [0. 0. 0.] degrees",
+        "  • family: sensor ",  # INVISIBLE SPACE
         "  • pixel: 15 ",  # INVISIBLE SPACE
     ]
     desc = s1.describe(return_string=True)
@@ -469,6 +472,7 @@ def test_describe():
         + "  • parent: None \n"
         + "  • position: [0. 0. 0.] mm\n"
         + "  • orientation: [0. 0. 0.] degrees\n"
+        + "  • family: sensor \n"
         + "  • pixel: 1 \n"
         + "  • style: SensorStyle(arrows=ArrowCS(x=ArrowSingle(color=None, show=True), "
         + "y=ArrowSingle(color=None, show=True), z=ArrowSingle(color=None, show=True)),"
@@ -489,6 +493,7 @@ def test_describe():
         + "  • parent: None \n"
         + "  • position: [0. 0. 0.] mm\n"
         + "  • orientation: [0. 0. 0.] degrees\n"
+        + "  • family: sensor \n"
         + "  • pixel: 1 \n"
         + "  • style: SensorStyle(arrows=ArrowCS(x=ArrowSingle(color=None, show=True), "
         + "y=ArrowSingle(color=None, show=True), z=ArrowSingle(color=None, show=True)),"
@@ -509,6 +514,7 @@ def test_describe():
         + "  • parent: None \n"
         + "  • position: [0. 0. 0.] mm\n"
         + "  • orientation: [0. 0. 0.] degrees\n"
+        + "  • family: sensor \n"
         + "  • pixel: 75 (3x5x5) "
     )
     desc = re.sub("id=*[0-9]*[0-9]", "id=REGEX", desc)
